+++ packages/fxa-auth-server/test/local/metrics/events.js
@@ -627,11 +627,7 @@ describe('metrics/events', () => {
           flow_id: 'bar',
           sync_device_count: 0,
           ua_browser: request.app.ua.browser,
-          ua_version: request.app.ua.browserVersion,
-          ua_os: request.app.ua.os,
-          user_country: 'United States',
-          user_locale: 'en-US',
-          user_state: 'California'
+          ua_version: request.app.ua.browserVersion
         }, 'log.amplitudeEvent was passed correct user properties')
 
         assert.equal(metricsContext.gather.callCount, 0, 'metricsContext.gather was not called')